Refactor: Enable decimal frequency and repetition values
This commit refactors the habit frequency and repetition per day settings to allow for decimal values.

Key changes:

- **Data Types:**
    - `HabitEntity.frequency` changed from `Int` to `Double`.
    - `AddHabitPageViewModel._habitFrequency` and `habitFrequency` changed from `Int?` to `Double?`.
    - `findHabitClusters` function now accepts `minOccurrences` as `Double` and returns `List<Triple<LocalDate, LocalDate, Double>>`.
    - `processDateTriples` function now accepts `List<Triple<LocalDate, LocalDate, Double>>`.
- **UI (FrequencySelector.kt):**
    - Input fields for frequency now accept and parse `Double` values.
    - Validation logic updated to handle `Double` inputs.
    - `isErrorDaily` state added for everyday frequency validation.
    - `KeyboardType.Number` applied to frequency text fields.
- **UI (RepetitionPerDaySelector.kt):**
    - `KeyboardType.Number` applied to the custom repetition text field.
    - `LaunchedEffect` added to update `viewModel.repetitionPerDay` when `finalRepetitionPerDay` changes.
- **ViewModel (AddHabitPageViewModel.kt):**
    - Added `_repetitionPerDay` and `repetitionPerDay` StateFlows to manage repetition per day.
    - `setRepetitionPerDay` function added.
    - Habit creation and update logic now includes `repetitionPerDay`.
    - `getHabitDetails` now populates `repetitionPerDay`.
- **Repository (HabitRepository.kt):**
    - `mapToHabit` now includes `repetitionPerDay`.
- **Minor formatting change** in `MonthCalendarRepository.kt`.

+++ app/src/main/java/com/zavedahmad/yaHabit/roomDatabase/HabitCompletionEntity.kt
@@ -21,7 +21,7 @@ data class HabitCompletionEntity(
     val habitId: Int,
     val completionDate: LocalDate,
     val partial : Boolean = false,
-    val repetitionsOnThisDay : Float = 1f,
+    val repetitionsOnThisDay :Double = 1.0,
     @PrimaryKey(autoGenerate = true) val id: Int = 0
 
 ): Parcelable
